+++ src/savant/view/swing/interval/BEDViewTrack.java
@@ -97,7 +97,7 @@ public class BEDViewTrack extends ViewTrack {
 
     public Resolution getDefaultModeResolution(Range range)
     {
-        int length = range.getLength();
+        long length = range.getLength();
 
         if (length < 10000) { return Resolution.VERY_HIGH; }
         else if (length < 50000) { return Resolution.HIGH; }
