+++ src/savant/view/swing/point/PointViewTrack.java
@@ -58,7 +58,7 @@ public class PointViewTrack extends ViewTrack {
 
     public Resolution getResolution(Range range)
     {
-        int length = range.getLength();
+        long length = range.getLength();
 
         if (length > 100000) { return Resolution.VERY_LOW; }
         return Resolution.VERY_HIGH;
